backend/pyproject.toml
[project]
name = "local-knowledge-base"
version = "0.1.0"
description = "Local hierarchical knowledge base with OpenAI-compatible LLM provider adapter"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pypdf>=4.3.0",
    "python-docx>=1.1.2",
]

[project.optional-dependencies]
embedding = [
    "sentence-transformers>=3.0.0",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
